# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/payment-links": {
      "get": {
        "operationId": "PaymentLinkController_getPaymentLinks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Payment links retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "links": [
                      {
                        "id": "lnk_123456",
                        "name": "Payment for Services",
                        "amount": 100,
                        "currency": "USD",
                        "url": "https://pay.paylinc.org/lnk_123456",
                        "status": "ACTIVE",
                        "createdAt": "2024-01-15T10:30:00Z"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user payment links",
        "tags": [
          "payment-links"
        ]
      },
      "post": {
        "operationId": "PaymentLinkController_createPaymentLink",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment link created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Payment link created successfully",
                    "link": {
                      "id": "lnk_123456",
                      "name": "Payment for Services",
                      "amount": 100,
                      "currency": "USD",
                      "url": "https://pay.paylinc.org/lnk_123456",
                      "status": "ACTIVE"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new payment link",
        "tags": [
          "payment-links"
        ]
      }
    },
    "/payment-links/{id}": {
      "get": {
        "operationId": "PaymentLinkController_getPaymentLinkDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment link details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "lnk_123456",
                    "name": "Payment for Services",
                    "amount": 100,
                    "currency": "USD",
                    "paymentMethods": {
                      "card": true,
                      "crypto": true
                    },
                    "createdBy": {
                      "name": "John Doe",
                      "email": "john@example.com"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get payment link details (public)",
        "tags": [
          "payment-links"
        ]
      },
      "patch": {
        "operationId": "PaymentLinkController_updatePaymentLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentLinkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment link updated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update payment link",
        "tags": [
          "payment-links"
        ]
      },
      "delete": {
        "operationId": "PaymentLinkController_deletePaymentLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment link deleted successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete payment link",
        "tags": [
          "payment-links"
        ]
      }
    },
    "/payment-links/{id}/initialize": {
      "post": {
        "operationId": "PaymentLinkController_initializePayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessPaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment initialized successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Payment initialized successfully",
                    "paymentLink": {
                      "id": "lnk_123456",
                      "name": "Payment for Services",
                      "amount": 100,
                      "currency": "USD"
                    },
                    "transactionId": "txn_123456"
                  }
                }
              }
            }
          }
        },
        "summary": "Initialize payment for frontend",
        "tags": [
          "payment-links"
        ]
      }
    },
    "/payment-links/{id}/confirm": {
      "post": {
        "operationId": "PaymentLinkController_confirmPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment confirmed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Payment confirmed successfully",
                    "transaction": {
                      "id": "txn_123456",
                      "status": "COMPLETED",
                      "amount": 100,
                      "currency": "USD"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Confirm payment completion",
        "tags": [
          "payment-links"
        ]
      }
    },
    "/payment-links/transactions/{id}": {
      "get": {
        "operationId": "PaymentLinkController_getTransactionDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction details retrieved successfully"
          },
          "404": {
            "description": "Transaction not found"
          }
        },
        "summary": "Get transaction details",
        "tags": [
          "payment-links"
        ]
      }
    },
...
  },
  "info": {
    "title": "EspeePay API",
    "description": "The EspeePay API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePaymentLinkDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name/description of the payment link",
            "example": "Payment for Design Services"
          },
          "amount": {
            "type": "number",
            "description": "Amount to be received",
            "example": 100,
            "minimum": 0.01
          },
          "currency": {
            "type": "string",
            "enum": [
              "USD",
              "GBP",
              "EUR",
              "NGN",
              "USDC",
              "USDT",
              "ESPEES"
            ],
            "description": "Currency for the payment",
            "example": "USD"
          }
        },
        "required": [
          "name",
          "amount",
          "currency"
        ]
      },
      "ProcessPaymentDto": {
        "type": "object",
        "properties": {
          "customerEmail": {
            "type": "string",
            "description": "Customer email address",
            "example": "customer@example.com"
          },
          "customerName": {
            "type": "string",
            "description": "Customer name",
            "example": "John Doe"
          },
          "paymentMethodToken": {
            "type": "string",
            "description": "Payment method token from Stripe/Paystack"
          },
          "cryptoAddress": {
            "type": "string",
            "description": "Crypto wallet address for crypto payments"
          }
        },
        "required": [
          "customerEmail",
          "customerName"
        ]
      },
      "UpdatePaymentLinkDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name/description of the payment link"
          },
          "amount": {
            "type": "number",
            "description": "Amount to be received",
            "minimum": 0.01
          },
          "currency": {
            "type": "string",
            "enum": [
              "USD",
              "GBP",
              "EUR",
              "NGN",
              "USDC",
              "USDT",
              "ESPEES"
            ],
            "description": "Currency for the payment"
          }
        }
      },
...
    }
  }
}